renamed auth to firebaseApp
